refactor: out 모듈 user 패키지 생성
